requirements/base.txt
aerich==0.6.3
aiosqlite==0.17.0
anyio==3.6.1
asgiref==3.5.2
asyncpg==0.25.0
cfgv==3.3.1
click==8.1.3
dictdiffer==0.9.0
distlib==0.3.4
fastapi==0.78.0
filelock==3.7.1
h11==0.13.0
httptools==0.4.0
identify==2.5.1
idna==3.3
iso8601==1.0.2
isort==5.10.1
loguru==0.6.0
mccabe==0.6.1
nodeenv==1.6.0
pathspec==0.9.0
platformdirs==2.5.2
pycodestyle==2.8.0
pydantic==1.9.1
pyflakes==2.4.0
pypika-tortoise==0.1.5
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
tortoise-orm==0.19.1
typing_extensions==4.2.0
uvicorn==0.17.6
uvloop==0.16.0
virtualenv==20.14.1
watchgod==0.8.2
websockets==10.3
